Initial Commit - very rough
THis is my first commit of oxbar, an OpenBSD x11 status bar... one of
those widgets that shows key system stats like battery power/status,
volume level, memory usage, cpu, etc. in a concise little bar on an X
display, which can be configured/shown outside of a window manager's tab
order or control.

This is very rough at this point. Currently it collects and shows the
following stats, in this order:
        1. battery status, plugged-in or on-battery, %remiaining, and
        time remaining.
        2. volume percent
        3. # of processes running
        4. memory information (amount active, total, and free)
        5. per cpu idle percents

The stats collection model is mostly done (needs some cleaning) and the
core X setup is also mostly done.

What's remaining is a great deal of polishing...
        1. Rendering of historical charts for memory & cpu
        2. Images/icons for stat names in the bar
        3. Cleaning up stats/* objects
        4. User specified configuration (everything's hardocded now)
        5. Overall refactoring, especially in oxbar.c
        6. More that i'm forgetting right now

Committing this now as it's in a usable state and I want to start
tracking history.

Note this is largely based on my xstatbar project, converting that to
use xcb and cairo/pango.