Update README to include bit about image support
This is low on my priority, but high on others. Being able to show
icons/images in place of "VOLUME", "MEMORY", etc is definitely more
concise and, I'll agree, visually appealing. That said, I just don't
care. I'm ok with text. I've added it to the list, but unless somebody
else contributes, this will likely be the last part I contribute.