BugFix: vertical backgrounds would span display
In a previous refactoring I introduced a bug where vertical bars (used
for volume and battery) would have their "red" portions of their bars
span the width of the display.

This corrects that bug and likely introduces many more.

It highlights the needs to further refactor the xcore/ui logic to better
decouple the two.