Refactor ui.* into xdraw.*
My original ui.* component was really just drawing primitives, and not
something that encompassed the whole "ui".

There's a need for both. Since the drawing primitives (drawing lines,
text, histograms, and basic double buffering) are essential and what was
implemented, I've renamed ui.* to xdraw.*, both the files and the APIs.

There are still some ui_* api's left (create and init). Those are going
to be moved next, into a new ui "proper" component that i'm chewing on.