Refactor all widget rendering to ui.* component
A large refactor, far from comlpete, but it's working.
This is the first step towards offloading all rendering of widgets to
the ui.* component. It introduces a ui_widget_foo_render(...) method for
every widget.

Much still to do - e.g. the widget rendering is tightly coupled to
stats. That's unavoidable, but i mixed approaches where all inputs to
the render methods are explict (listed-out, one-by-one) vs just passing
the global stats object.

I'll be chewing on that more and updating shortly, but this is a good
first step.