Further refactoring of ui.* components & oxbar.c
After last commit and chewing on idea, I decided to commit to coupling
the rendering of widgets with the stat components. Specifically, each
ui_widget_foo_render(...) method will now take at *least* two arguments:
the ui object, and one or more stat objects it wishes to render.

As of right now, there's a 1-1 correlation between stat objects and
widget renderers. That's not intentional and isn't something that should
be enforced -- consider for example having a stats object to collect
temperature sensors of cpus, and then having a cpu widget which shows
histograms & idle as well as temperatures. That render method would take
two stat objects: the cpu object and the temperature sensor object.

This nicely cleans the interface and, imho, makes the coupling clearer.
When updating a stat, you only have to modify the corresponding render
methods using that stat and NOT their interface and callout from
oxbar.c. That's an improvement.