Attempt at profiling using gprof
This adds a new target 'profile' in base that will attempt a normal
gprof buiild, so I can profile.

It doesn't work. clang used to not support -pg/gprof for ages, but does
now. There's a bug with PIE, so I'm using -fno-pie, but I suspect all
the additional archives i'm loading in gui/* and event stats/* aren't
supported. Bleh. I'm just adding this target and pushing to show it's of
interest.