Initial stab at abstracting settings
This is a first step towards supporting better configuration options.
This change creates a new settings object with default initializer, and
integrates it into the gui.* component.

It's not great. Actually it sucks.
It adds a great deal of complexity to the gui.* widget renderers, which
i'd rather avoid.

As such, I've stopped after the first widget. I have this in for now to
chew on further.