Redraw on EXPOSE events
This adds support for oxbar to instantly redraw whenever an expose event
is recieved (whenever the window is revealed, first mapped, etc). This
used to be annoying when returning from a full screen application (or
after killing an overlapping instance of oxbar while debugging!), and
having to wait up-to 1 second for oxbar to redraw (since it was
previously sleeping for 1 second in the main loop).

This works by separating the draw and update logic.
A timer is created to run every one second (using setitimer(2)) which
will refresh stats and redraw oxbar. The "main" loop in main() itself is
now a loop blocking on waiting for xcb events, which we only listen to
the expose event for now. When an expose event is recieved, we redraw
oxbar (no stats update).