Refactor BUT ABANDON this settings abstraction
This was a next attempt at creating a new settings abstraction.
It's unused outside the settings.* component apart from a print method
called in the beginning of oxbar.c.

It's bad -- nice clever use of macros to ease some pain, but looking up
settings after this would all be array searces doing string comparisons
:(

I'm committing an pushing out to store the work but will revert and
abandon this afterwards, in favor of a different method.