New Settings Abstraction
I removed the start of the abstraction introduced in the previous
commit, per the reasons outlined in that commit msg. This more-or-less
completes the previous attempt. I'm not as offended by it as I was
previously.

What remains is connecting this to command line flags and a config file.
That's next.

This effort really highlights for me the need to break widgets into
their own components. It makes sense.
A widget has it's own settings, it's own default logic, fallback, etc.
Ideally a widget just needs to work on a standard X hook with some
settings, and take in some stats, and that's it.

After my xdraw/gui refactoring (coming after settings) that will be
next.