UI Refactoring to ease widget drawing
This introduces a new gui/xdraw.* API that has a 'context' which tracks
WHAT you're drawing to (which says static) as well as WHERE you're
drawing (which changes as the rendering of the bar progresses).

The context tracks advancing x/y offset and greatly eases the logic in
the individual widget components in gui.*, making them easier to break
apart later.

It's also just awesome, and follows a pattern familiar to most
ui/graphics APIs (i'm not saying that's good...but it's familiar!)

More coming. My goal is to remove and isolate most of the X/cairo/pango
logic out of gui.* as much as possible.

I need to create new primitives in xdraw.* itself to better model and
encapsulate what the various widgets do. My goal is this: gui.* widgets
(and later independent widgets) *never* touch core x resources -- that's
all handled through xdraw.*. Histrograms/Time Series (poorly named) both
do this currently. That will be removed shortly.

Also this paves the way for a vertically stacked bar -- a side bar!