Bug Squashed: fix hang with concurrent drawing
SO - the past couple refactors were removing the old signal based timer
and switching to threads, primarily because the signal handler was
calling non-reentrant code (stats_update() and ui_draw()). That was done
as a cheap hack for the short term, and worked fine until I added the
xcb blocking event loop to handle expose events. Then occasionally it
would hang (every couple hours or so).

Two commits ago refactored that logic into threads and had that model
working, but draw could still be called concurrently. Since the
underlying cairo and pango apis aren't thread safe, there was still a
possibility for concurrent calls to clobber each other.

This commit guards that underlying logic with a mutex to prevent
concurrent calls to the gui api (that's also why i moved more of that
logic to xcore.*).

Bug fixed!