Fix bug with canceling/joining threads
This adds pthread_setcancelstate() to thread runners to better
coordinate the canceling of threads. That could causee an occasional
core on exit.

It also reorders the pthread_cancel calls to ensure the calling thread
(the signal handler runner) is the last to be cancelled.

Also removes an unused xcore.* method (xcore_destroy() ... too lazy to
sepaerate the diffs after failing two games pandemic legacy :(