New, more generic charting
This adds gui/chart.*, which intends to replace histogram and tseries.
It migrates the memory and cpu widgets to using it. Fun times this one.

The goal is to ditch histogram & tseries with the next commit, having a
single consolidated charting xdraw method, xdraw_chart(), that can
handle both percents and simple series.