Handle SIGTSTP / Suspend / ctrl-z
This just suspends the stats update thread when that signal is recieved
and resumes it on SIGCONT. No special handling of granularity of timing
between updates in those cases (not worthwhile imho).

This doesn't suspsend the gui thread - so it still responds to expose
events. That's tbd but in my case (and the common one, where this is
outside the wm's normal area of control) that's really rare.

It also doesn't detach from the terminal.