Create a global OXSTATS object wrapping the rest
This creates a common interface for all widgets by creating a common
stats interface that encompasses ALL stats (battery, volume, memory,
cpus, widgets, etc). That global object just connects to the existing
global created by each stats/* component (maybe not ideal, but to be
addressed later).

This may seem hacky, but I always wanted to decouple widgets from
specific stats/* components. E.g. it should be possible to create a
widget that combines memory and cpu info for a nice, tight display. In
the previous setup, that was doable by simply psasing both global stat
objects to such a renderer BUT that makes abstracting widget interfaces
more difficult. That abstraction will be necessary once I wish to allow
customization of which widgets are shown, in what order, and in what
alignment (left or right). Hence the abstraction. It's no real overhead
since I'm connecting to the existing abstractions with a static global
(which again, if necessary/desired I can decouple that further later).

In my mind, ideally widget rendering should take 3 components:
1. the CONTEXT on which to render (which also defines alignment)
2. the SETTINGS for that widget, and ideally ONLY that widget (so
   settings passed to a widget *should* be specific to that widget,
   which they aren't currently)
3. the STATS state -- all stats. Do what it wants with them.

This is a step towards that goal.