New Widget Abstraction & Usage
This creates a consistent abstraction for all widgets and demonstrates
how we can configure it going forwward. There are pros and cons, but
mostly pros.

It creates a new function type called widget_t and ensures all widgets
implement that interface. It defines two arrays of such widgets, one for
left-aligned, another for right-aligned (only time for now). The
gui_draw() method is greatly simplified by just looping through each and
calling the new widget_render() method.

The interface improvements are GREAT.

There are two BAD outcomes, where more boilerplate code is needed in
each widget::

First, each widget now has to draw their own hline / header line
above it. That's necessary because I can't determine the settings name
(settings->FOO->hdcolor) outside the widget with this abstraction. The
x offsets are easy, but I can't get the color name, and need more
tooling if I want to model that.

Second, I have to push "are the relevants stats objects enabledf?" logic
*out* of gui_draw() and into each individual widget method, and each now
returns true/false depending on if it was enabled (and thus rendered) or
not. That's more boilerplate.

Concerning both of these negatives....

1. I agree that BOTH points should live OUTSIDE gui.*
2. That logic *should* be specific to each widget

So to be clear: i'm not opposed to move that logic around. I'm only
opposed to adding more boilerplating. I need a better widget
interface/model.