Add gprof2dot visualization to gprof output
This is great - longer runs help prune the tree generated down.
Run for ~1-2 seconds to profile startup and shutdown (i don't really
care about that).
Run for more than a minute (at least) to profile the guts, and the chart
generated in gprof.png is awesomely useful.

Couple of insights so far: hex2rgba is costly and called EVERYWHERE.
I knew that. Total time is still pretty cheap, but it adds up. A cache
there or, better, resolving those colors everytime a color string is
set/changed would eliminate that cose.

Querying the battery is also costly (?). Need to check that out.

I'm stoked :D