Adding 1 hour diagram of gprof usage
Letting it run for an hour, gprof plus gprof2dot filter out all the big
stuff and really highlight where the time is spent. The first couple
levels aren't surprising, but the counts are (you don't see the counts
in this pic).

I'm adding a snapshot to the repo because it's...cool.