CPU Widget: Add new spin state and tweak colors
First - there's a new CPU states called "spin", and this adds support
for it. I'm still not 100% clear on what it represents.

I also reordered the drawing of the various states in the chart to be
more logical (imho) and improved the colors...solarized theme doesn't
work well for that many states and trying to distinguish between them
(not enough contrast).