Massive UI & Widget Refactoring
This work is far from complete - it eliminates some hacks and introduces
fresh new ones. Here are the key features it supports:

   * Proper left/right/center justified widgets
   * The defult display shows widgets using all three justifications

Under the hood, here's what happened:

   * ALL widget logic broken into a new component widgets.*
   * gui.* component now just orchestrates widgets
   * Each widget draws onto it's own, fresh context w/ surface/cairo
     objects
   * Those widgets are then mapped onto other contexts
   * xdraw.* cleaned-up to be "free" from xcore.* component (aside from
     init) - further decoupling the two
   * Widgets now have structure - their local state can be maintained
     outside gui...this will make creating/tweaking those easier in the
     future.

This is a great step toward better decoupling of logic.
Order of widgets (and now justification) is still hardcoded, but this is
a big step towards making them more customizable. Now that they can have
local state, along with setup/teardown, that will be easier to manage
going forward.

Still more refactoring & cleanup needed as part of this effort, but I'm
getting this out now since its functional and big step towards the goal.