stats/cpu.* NCPU -> NCPUONLINE
The sad state of intel. F@$%! that company and their poor practices for
years and years and years.

This now queries # of cpus based on those which are 'online' (actually
used), which is reduced thanks to Intel shit-headery. This is a new
interface. Works with the latest OpenBSD snapshots and prior.

On the plus side, there's more visual space in oxbar now that half my
cpus are gone! silver lining!