Make widgets configurable outside gui.* component
It's not used yet, but this now makes the gui.* component have the
widget configurations externally configurable.

Specifically the gui now tracks 3 separate lists of widgets: one for
left-aligned widgets, one for right-aligned, and one for centered.
It then exposes a method to add widgets to each context.

It configures a default that looks horrible but really tests the setup
:D

Up next is to plug that configuration into settings.* so that clients
can actually configure it.