New widgets.* api to create widgets from a string
This creates a new API in widgets.* component to create a list of
widgets in a space or comma separated string. It adds them to a
directional buffer (L2R, R2L, CENTERED) specified.

It's used in the existing widgets_init() method, replacing the old
hardcoding. The goal is to next expose this in settings, so that
widgets, and their placement, can be specified dynamically.