Refactor: Widgets & Placement now configurable
You can now use -S display.LeftWidgets/RightWidgets/CenterWidgets to set
the widgets shown in each portion of the display. The list is a space or
comma seperated list of widget names.

This moves more of the widget creation logic internally to widgets.*
component, removing it from settings.

After this comment, the general pattern is:

1. settings: load settings (defaults, command line, file)
2. create widgets based on settings.
...oxbar loop...
z. free() widgets created in step 2