Add configurable widget_spacing to gui
This re-adds support for a configurable "widget_spacing" property to the
gui, which represents the number of pixels to place between each widget
when painting them.

Note this is a settings specific to a gui, that gets copied over from
settings.* component on init.

This also corrects an outstanding bug I had with settings supporting -S
key=value on the command line, where I wasn't properly comparing strings
until now.