New feature: widget_bgcolor
This adds a new gui-level property called 'widget_bgcolor'.
It's applied as a background of every widget.

This is usefull when having a transparent background and you'd like the
widgets to be slightly-more opaque (or completely opaque).

It also permits rendering each widget as essentially their own
component, making the bar completely transparent and just having a
widget_bgcolor.