Remove padding as a xcore property
Similar to last commit, remove a gui-specific property from xcore.
It's still used in the init and setup of the height of the window,
injected by gui.* into xcore_init(), similar to how the last commit
injects bgcolor into xcore_clear(). That's still not ideal. I'd like to
remove all such logic from xcore.* and keep that just xcb/pango/cairo,
static, with no state. padding/bgcolor should be moved to xdraw.* imho.

This is a step in that direction.