Add gource target to Makefile
This is what happens when i get bored and look for a tool to track lines
of code per fine in a git repo over time....i find something else that's
nothing but eyecandy BUT SO FREAKIN SWEET!

Totally adding this as a target to track my awesome. oh yeah.