Refactor how widgets are set
Previously, setting the widgets shown in the left/center/right stacks
required the use of three different command line flags and settings (-l
widgets -c widgets -r widgets).

With this, those options are replaced by a single option -W widgets
where widgets is a string containing a space or comma seperated list of
widgets. They start in the left stack and after a vertical bar | are
then placed in the center stack, and after another bar are placed in the
right stack.