UI Refactor: breakup xcore.*
It's all still in the same code module but there are now 3 different
components all closely related:
1. the font (pango wrapper)
2. the x display information (simple xcb wrapper)
3. the window (wraps xcb + cairo)

This allows moving that damn padding logic out of xcore.* and into
oxbar. Not sure it should live there or in gui.*. But it definitely
didn't belong in xcore. This also greatly breaks-up/simplifies some
xcore logic and will help when i make the emmh properties configurable.