Refacor settings parsing again + support themes
This refactors the settings parsing again, to better support quoted
values like:
        display.font = "DejaVu Sans 16px"

Where, previously, that value of "DejaVu ..." was quoted and not
properly parsed.

It also adds supports for themes in the configuration file.

And we update the usage information to better document that the font
size should *always* be explicitly specified in pixels/px.