Reload config file on SIGHUP
On receiving a SIGHUP, oxbar will now reload the configuration file and
even respect any theme specified on the command line. There's one issue
though: if you change a line dealing with the widgets shown
(display.widgets), that's not reflected, because that configuration is
actually done outside the settings.* component.

I added a TODO to address that