SIGHUP reloads config even for display parameters
Now on a sighup, we not only reload the config, but afterwards we
totally destroy the gui and rebuild it. That means any structural
changes to display settings (like window placement or widgets) are also
reflected - that wasn't the case previously.

This uncovered a lot of bugs in the xcore library around cleanup, so
there's quite a few fixes there also.

I've added a new test case in testruns.sh to make sure this is tested
going forward.