Proper margin & padding support
The main window and each widget now support internal padding, with
seperate specs for the top, right, bottom, and left paddings (in
pixels).

The padding for the oxbar window itself is called the "margin" (since it
seperates the widgets from the window edge).

The padding internal to each widget is called "padding".

Padding is the same for each widget - you can't tweak padding per
widget (I think that's preferable, at least for now).

To set the padding, you can use the command line flags in the following
way:
1. -p sets the widget padding values
2. -m sets the margins (the padding for the oxbar window iteself)
3. Eacn can take as single number and set all padding components to
   that, or a sequence of 4 numbers.