Refactor xctx interface and move header into xdraw
This slightly refactors the xctx interface, specifically when creating a
context. Rather than a generic 'init' method, there are two: one makes a
root context and the other a scratchpad. I like that - but it could go
further.

This also moves the headerline logic completely out of gui and into
xdraw.