Network stats collector uses 'egress' interface
Previously I had hardcoded 'iwm0' as the interface to query for network
info in the stats/net.* collector. This changes that to instead use the
first interface in the 'egress' group, which will work for the vast
majority of laptop use cases (and likely desktop) and pick the
reasonable default.