Decouple widget implementations from most assets
This further decouples widget implementations from other parts of the
system and each other. Previous design had them behind a common
interface that imposed all the features that the most feature-rich
widget requried. This "dumbs it down" and puts the responsibility for
everything entirely within the widgets themselves.

Pros:
- all of the above
- new widgets can be onboarded with little/no understanding of other
  parts of the system
Cons:
- more boilerplate code is needed for each widget