Man page is nearly complete
There's been little details in the past few commits. Mostly because the
progress in the man page has been all over the place, and can only be
described as "progress".

This complets almost all sections. A further review is required, and
then the final section which includes the configurable properties for
all widgets needs to be done.