Progress on supporting non-px fonts
This is a step towards supporting fonts that specify sizes in something
other than pixels. Turns out there's a standard to translate all other
sizes to pixels and pango does that natively. I need to generate a
surface to draw on, however, to measure that size. That's the only way.

That itself is rather unpalatable, but in addition, there seems to be
some state or side effects of cairo/xcb when creating xcb backed
surfaces.

This approach I'm comitting here is not ideal for either of those
reasons, and can still segfault on relats when using non pixel based
fonts. oxbar would just error on startup previously though, so hey, it's
progress!