Refactor all stats/* components
This refactoring makes the stats components follow a pattern where:
1. they only export stats that are display-able (internal state is kept
   as private static locals in the module.c files)
2. the _init/_update/_cleanup() methods for each stats collector takes a
   struct _stats object of its type, rather than relying on a global
3. that passed struct is allocated and managed by the stats objects, so
   collectors don't have to worry about allocating/freeing that bit.
4. all stats collectors now cleanup as appropriate.

This also made some wide-ranging changes in widgets due to the
refactoring, and I took this opportunity to cleanup the inconsistent
naming of cpus/cpushort/cpuslong widgets (the prefix is now "cpu"), and
updated corresponding callers, tests, and screenshots.

Regen'd TODO and loc with this.