Makefile refactor: move objects to their own include
This didn't accomplish what i want :(
I was hoping to have the objects/dependencies of each subdir defined in
some common include that each subdir/Makefile could include, as well as
the root Makefile. That doesn't work though - the root Makefile needs
the "subdir/" prefix on each object name for the recursion to work.

Ugh. This keeps bothering me. I have a working solution now and will not
break that. This is a small step towards what I want though (each subdir
defining its own list of objects, which it uses, and the root Makefile
uses that).